Adapted last merge
