+++ src/main/kotlin/com/buttersus/wiremaster/screen/ModConfigScreen.kt
@@ -1,7 +1,7 @@
-package com.buttersus.wireworks.screen
+package com.buttersus.wiremaster.screen
 
-import com.buttersus.wireworks.WireWorks
-import com.buttersus.wireworks.config.ConfigManager
+import com.buttersus.wiremaster.WireMaster
+import com.buttersus.wiremaster.config.ConfigManager
 import me.shedaniel.clothconfig2.api.ConfigBuilder
 import me.shedaniel.clothconfig2.api.ConfigCategory
 import me.shedaniel.clothconfig2.api.ConfigEntryBuilder
@@ -13,21 +13,21 @@ import net.minecraft.text.Text
 @Environment(EnvType.CLIENT)
 class ModConfigScreen(
     parent: Screen?
-) : Screen(Text.translatable("config.${WireWorks.MOD_ID}.title")) {
+) : Screen(Text.translatable("config.${WireMaster.MOD_ID}.title")) {
     private val configScreen: Screen
 
     init {
         val builder = ConfigBuilder.create()
             .setParentScreen(parent)
-            .setTitle(Text.translatable("config.${WireWorks.MOD_ID}.title"))
+            .setTitle(Text.translatable("config.${WireMaster.MOD_ID}.title"))
 
-        val general: ConfigCategory = builder.getOrCreateCategory(Text.translatable("config.${WireWorks.MOD_ID}.general"))
+        val general: ConfigCategory = builder.getOrCreateCategory(Text.translatable("config.${WireMaster.MOD_ID}.general"))
         general.addEntry(
             ConfigEntryBuilder.create()
-                .startBooleanToggle(Text.translatable("config.${WireWorks.MOD_ID}.experimental_orthographic"), WireWorks.EXPERIMENTAL_ORTHOGRAPHIC)
+                .startBooleanToggle(Text.translatable("config.${WireMaster.MOD_ID}.experimental_orthographic"), WireMaster.EXPERIMENTAL_ORTHOGRAPHIC)
                 .setDefaultValue(false)
-                .setTooltip(Text.translatable("config.${WireWorks.MOD_ID}.experimental_orthographic.tooltip"))
-                .setSaveConsumer { WireWorks.EXPERIMENTAL_ORTHOGRAPHIC = it }
+                .setTooltip(Text.translatable("config.${WireMaster.MOD_ID}.experimental_orthographic.tooltip"))
+                .setSaveConsumer { WireMaster.EXPERIMENTAL_ORTHOGRAPHIC = it }
                 .build()
         )
