+++ src/main/resources/assets/wire-master/lang/en_us.json
@@ -1,4 +1,8 @@
 {
   "category.wire-master.keybindings": "Wire Master",
-  "key.wire-master.toggle_wire_designer": "Toggle wire designer"
+  "key.wire-master.toggle_wire_designer": "Toggle wire designer",
+  "config.wire-master.title": "Wire Master configuration",
+  "config.wire-master.general": "General",
+  "config.wire-master.experimental_orthographic": "Experimental orthographic view",
+  "config.wire-master.experimental_orthographic.tooltip": "Enable experimental orthographic projection (not recommended)"
 }
