+++ src/main/resources/assets/wire-master/lang/ru_ru.json
@@ -1,4 +1,8 @@
 {
   "category.wire-master.keybindings": "Wire Master",
-  "key.wire-master.toggle_wire_designer": "Переключить редактор проводов"
+  "key.wire-master.toggle_wire_designer": "Переключить wire designer",
+  "config.wire-master.title": "Настройки Wire Master",
+  "config.wire-master.general": "Основное",
+  "config.wire-master.experimental_orthographic": "Экспериментальная ортографическое отображение",
+  "config.wire-master.experimental_orthographic.tooltip": "Включить экспериментальное ортографическое отображение (не рекомендуется)"
 }
